Swapping over to old controller
