+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/UploadDocumentsController.java
@@ -11,10 +11,13 @@ import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
+import uk.gov.hmcts.reform.fpl.service.ApplicationDocumentsService;
 import uk.gov.hmcts.reform.fpl.service.DocumentsValidatorService;
 import uk.gov.hmcts.reform.fpl.service.UploadDocumentsService;
 
 import java.util.List;
+
+import static uk.gov.hmcts.reform.fpl.enums.YesNo.YES;
 
 @Api
 @RestController
@@ -23,6 +26,16 @@ import java.util.List;
 public class UploadDocumentsController extends CallbackController {
     private final DocumentsValidatorService documentsValidatorService;
     private final UploadDocumentsService uploadDocumentsService;
+    private final ApplicationDocumentsService applicationDocumentsService;
+
+    @PostMapping("/about-to-start")
+    public AboutToStartOrSubmitCallbackResponse handleAboutToStart(@RequestBody CallbackRequest callbackrequest) {
+        CaseDetails caseDetails = callbackrequest.getCaseDetails();
+
+        caseDetails.getData().remove("showMetaFields");
+
+        return respond(caseDetails);
+    }
 
     @PostMapping("/mid-event")
     public AboutToStartOrSubmitCallbackResponse handleMidEvent(@RequestBody CallbackRequest callbackrequest) {
@@ -37,5 +50,19 @@ public class UploadDocumentsController extends CallbackController {
         }
 
         return respond(caseDetails, errors);
+    }
+
+    @PostMapping("/about-to-submit")
+    public AboutToStartOrSubmitCallbackResponse handleAboutToSubmit(@RequestBody CallbackRequest callbackrequest) {
+        CaseDetails caseDetails = callbackrequest.getCaseDetails();
+        CaseData caseData = getCaseData(caseDetails);
+
+        CaseData caseDataBefore = getCaseDataBefore(callbackrequest);
+        caseDetails.getData().putAll(applicationDocumentsService.updateCaseDocuments(caseData.getDocuments(),
+            caseDataBefore.getDocuments()));
+
+        caseDetails.getData().put("showMetaFields", YES);
+
+        return respond(caseDetails);
     }
 }
