+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/FeatureToggleService.java
@@ -68,7 +68,7 @@ public class FeatureToggleService {
     }
 
     public boolean isApplicationDocumentsEventEnabled() {
-        return ldClient.boolVariation("application-documents-event", createLDUser(), false);
+        return ldClient.boolVariation("application-documents-event", createLDUser(), true);
     }
 
     private LDUser createLDUser() {
